xu ly api thanh toan gio hang, shipping info

+++ database/migrations/2022_01_07_161425_create_order_details_table.php
@@ -18,7 +18,7 @@ class CreateOrderDetailsTable extends Migration
             $table->unsignedBigInteger('product_id');
             $table->unsignedBigInteger('order_id');
             $table->integer('quantity');
-            $table->float('unit_price');
+            $table->double('unit_price');
             $table->softDeletes();
             $table->timestamps();
         });
